correct bracket error causing tests to fail in Decoder logic
