+++ src/main/java/com/CaesarCipher/App.java
@@ -1,12 +1,10 @@
-//In the future this is where the Front-End logic will go
+package com.CaesarCipher;//In the future this is where the Front-End logic will go
+import java.util.Scanner;
+
 
 public class App {
 
-    public void main(String[] args){
 
-
-
-    }
