--- src/test/java/com/CaesarCipher/CaesarCipherTest.java
@@ -1,17 +1,0 @@
-
-package com.CaesarCipher;
-import org.junit.*;
-import static org.junit.Assert.*;
-import java.lang.*;
-
-public class CaesarCipherTest {
-    @Test
-    public void testMain(){
-
-
-    }
-
-
-
-
-}
